refactor: relocate "lib::types::Msg" and related to be in TUI
For now only the ones only used in the TUI.
