+++ tui/src/ui/components/config_editor/color.rs
@@ -26,7 +26,6 @@ use termusiclib::config::SharedTuiSettings;
 use termusiclib::config::v2::tui::theme::ThemeWrap;
 use termusiclib::config::v2::tui::theme::styles::ColorTermusic;
 use termusiclib::ids::{Id, IdConfigEditor};
-use termusiclib::types::{ConfigEditorMsg, Msg};
 use tui_realm_stdlib::{Label, Select, Table};
 use tuirealm::command::{Cmd, CmdResult, Direction, Position};
 use tuirealm::event::{Key, KeyEvent, KeyModifiers};
@@ -38,6 +37,7 @@ use tuirealm::{AttrValue, Attribute, Component, Event, MockComponent, State, Sta
 
 use crate::ui::components::vendored::tui_realm_stdlib_input::Input;
 use crate::ui::model::{Model, UserEvent};
+use crate::ui::msg::{ConfigEditorMsg, Msg};
 
 const COLOR_LIST: [ColorTermusic; 19] = [
     ColorTermusic::Reset,
