+++ tui/src/ui/components/popups/deleteconfirm.rs
@@ -1,6 +1,5 @@
 use termusiclib::config::{SharedTuiSettings, TuiOverlay};
 use termusiclib::ids::Id;
-use termusiclib::types::{DeleteConfirmMsg, Msg};
 use tuirealm::{
     Component, Event, MockComponent, State, StateValue,
     command::{Cmd, CmdResult, Direction, Position},
@@ -11,6 +10,7 @@ use tuirealm::{
 use super::{YNConfirm, YNConfirmStyle};
 use crate::ui::components::vendored::tui_realm_stdlib_input::Input;
 use crate::ui::model::{Model, UserEvent};
+use crate::ui::msg::{DeleteConfirmMsg, Msg};
 
 /// Component for a "Are you sure to delete? Y/N" popup
 ///
