+++ tui/src/ui/components/popups/general_search.rs
@@ -4,7 +4,6 @@ use anyhow::{Result, anyhow, bail};
 use termusiclib::config::{SharedTuiSettings, TuiOverlay};
 use termusiclib::ids::Id;
 use termusiclib::track::MediaTypes;
-use termusiclib::types::{GSMsg, Msg};
 use tui_realm_stdlib::Table;
 use tui_realm_treeview::TREE_INITIAL_NODE;
 use tuirealm::command::{Cmd, CmdResult, Direction, Position};
@@ -15,6 +14,7 @@ use tuirealm::{AttrValue, Attribute, Component, Event, MockComponent, State, Sta
 use crate::ui::Model;
 use crate::ui::components::vendored::tui_realm_stdlib_input::Input;
 use crate::ui::model::UserEvent;
+use crate::ui::msg::{GSMsg, Msg};
 
 #[derive(MockComponent)]
 pub struct GSInputPopup {
