+++ tui/src/ui/components/popups/youtube_search.rs
@@ -1,4 +1,3 @@
-use crate::ui::components::vendored::tui_realm_stdlib_input::Input;
 /**
  * MIT License
  *
@@ -22,15 +21,17 @@ use crate::ui::components::vendored::tui_realm_stdlib_input::Input;
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-use crate::ui::model::{Model, UserEvent};
 use termusiclib::config::{SharedTuiSettings, TuiOverlay};
 use termusiclib::ids::Id;
-use termusiclib::types::{Msg, YSMsg};
 use tui_realm_stdlib::Table;
 use tuirealm::command::{Cmd, CmdResult, Direction, Position};
 use tuirealm::event::{Key, KeyEvent, KeyModifiers};
 use tuirealm::props::{Alignment, BorderType, Borders, InputType, TableBuilder, TextSpan};
 use tuirealm::{Component, Event, MockComponent, State, StateValue};
+
+use crate::ui::components::vendored::tui_realm_stdlib_input::Input;
+use crate::ui::model::{Model, UserEvent};
+use crate::ui::msg::{Msg, YSMsg};
 
 #[derive(MockComponent)]
 pub struct YSInputPopup {
