+++ tui/src/ui/components/progress.rs
@@ -1,10 +1,10 @@
 use std::time::Duration;
 
+use crate::ui::msg::Msg;
 use termusiclib::config::TuiOverlay;
 use termusiclib::ids::Id;
 use termusiclib::track::DurationFmtShort;
 use termusiclib::track::MediaTypesSimple;
-use termusiclib::types::Msg;
 use tui_realm_stdlib::ProgressBar;
 use tuirealm::props::{Alignment, BorderType, Borders, PropPayload, PropValue};
 use tuirealm::{AttrValue, Attribute, Component, Event, MockComponent};
